fixed jwt stuff

+++ app/routers/comment_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from typing import List
-from app.crud import get_comments, create_comment, update_comment, delete_comment
+from app.crud import get_comment, get_comments, create_comment, update_comment, delete_comment
 from app.models.base import SessionLocal
 from app.schemas import Comment, CommentRead, CommentCreate, CommentUpdate, Token
 from jwt import verify_access_token
@@ -28,6 +28,8 @@ async def post_comment_endpoint(post_id: int,
                                 comment: CommentCreate, 
                                 session: AsyncSession = Depends(get_session),
                                 user: Token = Depends(verify_access_token)):
+    if owner_id != user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to post comment under different owner")
     created_comment = await create_comment(post_id, owner_id, comment, session)
     if created_comment is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Error creating comment")
@@ -35,19 +37,29 @@ async def post_comment_endpoint(post_id: int,
 
 @router.put("/comments/{comment_id}", response_model=CommentRead)
 async def update_comment_endpoint(comment_id: int,
-                                  comment: CommentUpdate,
+                                  new_comment: CommentUpdate,
                                   session: AsyncSession = Depends(get_session),
                                   user: Token = Depends(verify_access_token)):
-    updated_comment = await update_comment(session, comment_id, comment)
-    if update_comment is None:
+    comment = get_comment(session, comment_id)
+    if comment is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Comment not found")
+    if comment.owner_id != user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this comment")
+    updated_comment = await update_comment(session, comment_id, new_comment)
     return updated_comment
 
 @router.delete("/comments/{comment_id}", status_code=status.HTTP_200_OK)
 async def delete_comment_endpoint(comment_id: int, 
                                   session: AsyncSession = Depends(get_session),
                                   user: Token = Depends(verify_access_token)):
-    deleted_comment = await delete_comment(session, comment_id)
-    if not deleted_comment:
+    comment = get_comment(session, comment_id)
+
+    if comment is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Comment not found")
+    
+    if comment.owner_id != user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this comment")
+    
+    await delete_comment(session, comment_id)
+    
     return {"detail": "Comment deleted successfully"}
